removed PAI and _original from first PR.  moved _scheduler to replace the current file
